+++ frappe/Link/index.ts
@@ -1,2 +1,0 @@
-export { default as Link } from './Link.vue'
-export type { LinkEmits, LinkExposed, LinkOption, LinkProps } from './types'